fixed errors with entities auto generated
