--- src/main/java/com/sparta/zmsb/weekfiveteamproject/CountrylanguageEntity.java
@@ -1,66 +1,0 @@
-package com.sparta.zmsb.weekfiveteamproject;
-
-import com.sparta.zmsb.weekfiveteamproject.entities.CountryEntity;
-import com.sparta.zmsb.weekfiveteamproject.entities.CountrylanguageEntityId;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import org.hibernate.annotations.ColumnDefault;
-
-import java.math.BigDecimal;
-
-@Entity
-@Table(name = "countrylanguage", schema = "world")
-public class CountrylanguageEntity {
-    @EmbeddedId
-    private CountrylanguageEntityId id;
-
-    @MapsId("countryCode")
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @ColumnDefault("''")
-    @JoinColumn(name = "CountryCode", nullable = false)
-    private CountryEntity countryCode;
-
-    @NotNull
-    @ColumnDefault("'F'")
-    @Lob
-    @Column(name = "IsOfficial", nullable = false)
-    private String isOfficial;
-
-    @NotNull
-    @ColumnDefault("0.0")
-    @Column(name = "Percentage", nullable = false, precision = 4, scale = 1)
-    private BigDecimal percentage;
-
-    public CountrylanguageEntityId getId() {
-        return id;
-    }
-
-    public void setId(CountrylanguageEntityId id) {
-        this.id = id;
-    }
-
-    public CountryEntity getCountryCode() {
-        return countryCode;
-    }
-
-    public void setCountryCode(CountryEntity countryCode) {
-        this.countryCode = countryCode;
-    }
-
-    public String getIsOfficial() {
-        return isOfficial;
-    }
-
-    public void setIsOfficial(String isOfficial) {
-        this.isOfficial = isOfficial;
-    }
-
-    public BigDecimal getPercentage() {
-        return percentage;
-    }
-
-    public void setPercentage(BigDecimal percentage) {
-        this.percentage = percentage;
-    }
-
-}
